Move useStore to hooks file
